Renamed CastError constructor for BadRequestError; replaced ValidationError with BadRequestError; deleted ValidationError constructor
