+++ errors/bad-request-error.js
@@ -1,8 +1,8 @@
-class CastError extends Error {
+class BadRequestError extends Error {
   constructor(message) {
     super(message);
     this.statusCode = 400;
   }
 }
 
-module.exports = CastError;
+module.exports = BadRequestError;
